# swagger/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.1
  title: BankManager
  description: lol api
servers:
  - url: "http://notlocalhost:8080"
paths:
...
  /surname/{surname}:
    get:
      tags:
        - user
      description: This is a getdata call
      operationId: getBanksBySurname
      parameters:
        - name: surname
          in: path
          description: surname variable
          required: true
          style: simple

          schema:
            maxLength: 255
            type: string
      responses:
        '200':
          description: This is 200 description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JpaBankEntity'
...
components:
  schemas:
    JpaBankEntity:
      type: object
      properties:
        accountNumber:
          type: string
          example: sdwd3d32dd
        trust:
          type: number
          example: 3.14
        transactionFee:
          type: integer
          example: 42
        surname:
          type: string
          example: Doe
...